Fix tools build, binstall/xinstall, for systems without id_t, e.g. MacOS X 10.3.9.  This should better match the NetBSD definition of id_t in sys/sys/types.h.

+++ tools/compat/compat_defs.h
@@ -1,4 +1,4 @@
-/*	$NetBSD: compat_defs.h,v 1.57 2006/10/12 16:19:06 christos Exp $	*/
+/*	$NetBSD: compat_defs.h,v 1.58 2007/01/09 17:34:27 ginsbach Exp $	*/
 
 #ifndef	__NETBSD_COMPAT_DEFS_H__
 #define	__NETBSD_COMPAT_DEFS_H__
@@ -146,7 +146,7 @@ struct passwd;
 /* Type substitutes. */
 
 #if !HAVE_ID_T
-typedef unsigned long id_t;
+typedef unsigned int id_t;
 #endif
 
 #if !HAVE_SOCKLEN_T
